feat: add dsse verifier threshold
Adds the ability to require at least a specified number of verifiers to
successfully verify a DSSE envelope.

Also fixes witness issue #194 where multiple public keys within a policy
will cause all verification to fail. This was due to the DSSE verify
function returning as soon as a single verifier fails. Now the code
appropriately tries to verify against all passed in verifiers and only
throws an error if no verifier successfully verifies the envelope or the
threshold is unmet.

Signed-off-by: Mikhail Swift <mikhail@testifysec.com>