fix: reset verifier each iteration while loading pub keys from policy (#173)
* fix: reset verifier each iteration while loading pub keys from policy

Ran into an issue testing a policy with multiple public key
functionaries. I would continually get ErrKeyIDMismatch when loading the
second key. Turns out the verifier variable was being persisted on each
iteration and failing the nil check to reload the new verifier.
Resetting this within the for loop fixes it.

* fix: add material and product attestors to imports

Signed-off-by: Mikhail Swift <mikhail@testifysec.com>