feat: add secretscan attestor to detect secrets in products
The secretscan attestor provides a mechanism to detect secrets in products and attestations. It can be configured to scan binary files and to fail the attestation process if secrets are detected.

This is useful for users who want to ensure that no secrets are included in their products before they are released or to detect if any secrets were inadvertently included during the build process.

Signed-off-by: Cole Kennedy <cole@testifysec.com>