fix: use sha384 for signing proof of possession challenge from fulcio (#490)
By default we generate an elliptic P384 for our fulcio signer. When we
get the proof of possession challenge from fulcio, we were hashing it
with SHA256 during the signing process. This was working fine until
recently, when we started getting failures while getting certificates
from Fulcio.

This seems to trace back to this commit:
https://github.com/sigstore/sigstore/commit/63abeebb6c891fb583a556210cb6d593a95e688b

This adds a registry to map the type of public key and which algorithms
to use by default. This expects SHA384 to be used in our case.

Swapping the signer to use SHA384 when signing the challenge fixes our
error.

Signed-off-by: Mikhail Swift <mikhail@testifysec.com>